lib/tapi_tcp_states: ignore known faulty TCP state moves
If we go from TCP_FIN_WAIT2 or TCP_CLOSING to TCP_TIME_WAIT
getsockopt() says that we are in TCP_CLOSE. Also if we go from
TCP_LISTEN to TCP_SYN_RECV it says that we are in TCP_LISTEN.
There is a bug https://bugzilla.kernel.org/show_bug.cgi?id=33902
(with status RESOLVED WILL_NOT_FIX) that is closely related to the
first case of this problem.
It is a strange idea to check this TCP move if TCP check is faulty.

Signed-off-by: Nikolai Kosovskii <nikolai.kosovskii@arknetworks.am>